github.io: Add mini project deck.html
This is a project that I worked on in my CS316 class. This is a web game made using HTML, CSS and JavaScript where a deck of cards is fetched and 5 cards randomly are chosen and flipped backwards. The player then selects option black/red and in case the selected color is the higher number of colors in 5 sets, the player wins.